Refactored CoreProtectAPI
